+++ schemaParts/tester/src/java/pko/KiCadLogicalSchemeSimulator/components/tester/TestInBus.java
@@ -51,7 +51,7 @@ public class TestInBus extends SchemaPart {
             }
 
             @Override
-            public void setState(long newState) {
+            public void setState(int newState) {
                 hiImpedance = false;
                 state = newState;
             }
